Replace the slider and reorganize the theme to match vuepress 1.x's style
