+++ docs/.vuepress/theme/components/AlgoliaSearchBox.vue
@@ -1,65 +1,67 @@
 <template>
-  <form
-    id="search-form"
-    class="algolia-search-wrapper search-box"
-  >
-    <input
-      id="algolia-search-input"
-      class="search-query"
-    >
+  <form id="search-form" class="algolia-search-wrapper search-box">
+    <input id="algolia-search-input" class="search-query" />
   </form>
 </template>
 
 <script>
 export default {
-  props: ['options'],
+  props: ["options"],
 
-  mounted () {
-    this.initialize(this.options, this.$lang)
+  mounted() {
+    this.initialize(this.options, this.$lang);
   },
 
   methods: {
-    initialize (userOptions, lang) {
+    initialize(userOptions, lang) {
       Promise.all([
-        import(/* webpackChunkName: "docsearch" */ 'docsearch.js/dist/cdn/docsearch.min.js'),
-        import(/* webpackChunkName: "docsearch" */ 'docsearch.js/dist/cdn/docsearch.min.css')
+        import(
+          /* webpackChunkName: "docsearch" */ "docsearch.js/dist/cdn/docsearch.min.js"
+        ),
+        import(
+          /* webpackChunkName: "docsearch" */ "docsearch.js/dist/cdn/docsearch.min.css"
+        )
       ]).then(([docsearch]) => {
-        docsearch = docsearch.default
-        const { algoliaOptions = {}} = userOptions
-        docsearch(Object.assign(
-          {},
-          userOptions,
-          {
-            inputSelector: '#algolia-search-input',
+        docsearch = docsearch.default;
+        const { algoliaOptions = {} } = userOptions;
+        docsearch(
+          Object.assign({}, userOptions, {
+            inputSelector: "#algolia-search-input",
             // #697 Make docsearch work well at i18n mode.
-            algoliaOptions: Object.assign({
-              'facetFilters': [`lang:${lang}`].concat(algoliaOptions.facetFilters || [])
-            }, algoliaOptions)
-          }
-        ))
-      })
+            algoliaOptions: Object.assign(
+              {
+                facetFilters: [`lang:${lang}`].concat(
+                  algoliaOptions.facetFilters || []
+                )
+              },
+              algoliaOptions
+            )
+          })
+        );
+      });
     },
 
-    update (options, lang) {
-      this.$el.innerHTML = '<input id="algolia-search-input" class="search-query">'
-      this.initialize(options, lang)
+    update(options, lang) {
+      this.$el.innerHTML =
+        '<input id="algolia-search-input" class="search-query">';
+      this.initialize(options, lang);
     }
   },
 
   watch: {
-    $lang (newValue) {
-      this.update(this.options, newValue)
+    $lang(newValue) {
+      this.update(this.options, newValue);
     },
 
-    options (newValue) {
-      this.update(newValue, this.$lang)
+    options(newValue) {
+      this.update(newValue, this.$lang);
     }
   }
-}
+};
 </script>
 
 <style lang="stylus">
-@import './styles/config.styl'
+@import '../styles/config.styl'
 
 .algolia-search-wrapper
   & > span
@@ -152,5 +154,4 @@ export default {
       width 5px
       margin -3px 3px 0
       vertical-align middle
-
 </style>
