+++ docs/.vuepress/theme/components/DropdownLink.vue
@@ -77,7 +77,7 @@ export default {
 </script>
 
 <style lang="stylus">
-@import './styles/config.styl'
+@import '../styles/config.styl'
 
 .dropdown-wrapper
   cursor pointer
