+++ docs/.vuepress/theme/components/NavLink.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script>
-import { isExternal, isMailto, isTel, ensureExt } from './util'
+import { isExternal, isMailto, isTel, ensureExt } from '../util'
 
 export default {
   props: {
