+++ docs/.vuepress/theme/components/NavLinks.vue
@@ -35,7 +35,7 @@
 
 <script>
 import DropdownLink from './DropdownLink.vue'
-import { resolveNavLinkItem } from './util'
+import { resolveNavLinkItem } from '../util'
 import NavLink from './NavLink.vue'
 
 export default {
@@ -116,7 +116,7 @@ export default {
 </script>
 
 <style lang="stylus">
-@import './styles/config.styl'
+@import '../styles/config.styl'
 
 .nav-links
   display inline-block
