+++ docs/.vuepress/theme/components/Navbar.vue
@@ -100,7 +100,7 @@ function css(el, property) {
 </script>
 
 <style lang="stylus">
-@import './styles/config.styl'
+@import '../styles/config.styl'
 
 $navbar-vertical-padding = 0.7rem
 $navbar-horizontal-padding = 1.5rem
