+++ docs/.vuepress/theme/components/Page.vue
@@ -63,7 +63,7 @@
 </template>
 
 <script>
-import { resolvePage, normalize, outboundRE, endingSlashRE } from './util'
+import { resolvePage, normalize, outboundRE, endingSlashRE } from '../util'
 
 export default {
   props: ['sidebarItems'],
@@ -197,8 +197,8 @@ function find (page, items, offset) {
 </script>
 
 <style lang="stylus">
-@import './styles/config.styl'
-@require './styles/wrapper.styl'
+@import '../styles/config.styl'
+@require '../styles/wrapper.styl'
 
 .page
   padding-bottom 2rem
