+++ docs/.vuepress/theme/components/SidebarLink.vue
@@ -1,5 +1,5 @@
 <script>
-import { isActive, hashRE, groupHeaders } from './util'
+import { isActive, hashRE, groupHeaders } from '../util'
 
 export default {
   functional: true,
@@ -60,7 +60,7 @@ function renderChildren (h, children, path, route, maxDepth, depth = 1) {
 </script>
 
 <style lang="stylus">
-@import './styles/config.styl'
+@import '../styles/config.styl'
 
 .sidebar .sidebar-sub-headers
   padding-left 1rem
