+++ docs/.vuepress/theme/styles/theme.styl
@@ -191,4 +191,4 @@ th, td
       padding-left 0
 
 @require './mobile.styl'
-@require '../../style.styl'
+@require './custom-style.styl'
